requirements.txt
audioread==3.0.1
braceexpand==0.1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
h5py==3.12.1
huggingface_hub==0.25.2
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
matplotlib==3.7.1
msgpack==1.1.0
numba==0.60.0
numpy==1.26.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
progressbar==2.5
protobuf==5.27.2
psutil==6.0.0
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
RPi.GPIO==0.7.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.14.0
sentry-sdk==2.7.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==1.11.0
torchaudio==0.11.0
torchlibrosa==0.1.0
torchvision==0.12.0
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.3
wcwidth==0.2.13
webdataset==0.2.86
wget==3.2
